backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
groq>=0.9.0
langgraph>=0.2.16

pytest>=8.2.0
httpx>=0.27.0
